requirements.txt
anyio==4.10.0
certifi==2025.8.3
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
requests==2.32.5
fastapi==0.118.0
uvicorn==0.37.0
python-multipart